+++ torchdata/utils/__init__.py
@@ -1,0 +1,11 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from torchdata.utils.pin_memory import pin_memory_fn
+
+__all__ = ["pin_memory_fn"]
+
+assert __all__ == sorted(__all__)
